Move expandFilePath to pkg/action
